[Refactor] successResponseDocument 메서드로 분리
