+++ src/test/java/com/inq/wishhair/wesharewishhair/auth/controller/utils/LoginRequestUtils.java
@@ -4,8 +4,8 @@ import com.inq.wishhair.wesharewishhair.auth.controller.dto.request.LoginRequest
 
 public abstract class LoginRequestUtils {
 
-    private static final String loginId = "loginId";
-    private static final String pw = "pw";
+    private static final String loginId = "hello12@naver.com";
+    private static final String pw = "hello1234!";
 
     public static LoginRequest createRequest() {
         return new LoginRequest(loginId, pw);
