+++ src/test/java/com/inq/wishhair/wesharewishhair/global/fixture/PointFixture.java
@@ -1,4 +1,4 @@
-package com.inq.wishhair.wesharewishhair.fixture;
+package com.inq.wishhair.wesharewishhair.global.fixture;
 
 import com.inq.wishhair.wesharewishhair.user.domain.User;
 import com.inq.wishhair.wesharewishhair.user.domain.point.PointHistory;
