+++ src/test/java/com/inq/wishhair/wesharewishhair/hairstyle/service/HairStyleSearchServiceTest.java
@@ -1,6 +1,6 @@
 package com.inq.wishhair.wesharewishhair.hairstyle.service;
 
-import com.inq.wishhair.wesharewishhair.fixture.UserFixture;
+import com.inq.wishhair.wesharewishhair.global.fixture.UserFixture;
 import com.inq.wishhair.wesharewishhair.global.base.ServiceTest;
 import com.inq.wishhair.wesharewishhair.global.dto.response.ResponseWrapper;
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
@@ -15,7 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 
-import static com.inq.wishhair.wesharewishhair.fixture.HairStyleFixture.*;
+import static com.inq.wishhair.wesharewishhair.global.fixture.HairStyleFixture.*;
 import static com.inq.wishhair.wesharewishhair.global.utils.PageableUtils.getDefaultPageable;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
