+++ src/test/java/com/inq/wishhair/wesharewishhair/review/controller/ReviewControllerTest.java
@@ -1,12 +1,10 @@
 package com.inq.wishhair.wesharewishhair.review.controller;
 
-import com.inq.wishhair.wesharewishhair.fixture.ReviewFixture;
+import com.inq.wishhair.wesharewishhair.global.fixture.ReviewFixture;
 import com.inq.wishhair.wesharewishhair.global.base.ControllerTest;
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
 import com.inq.wishhair.wesharewishhair.review.controller.dto.request.ReviewCreateRequest;
 import com.inq.wishhair.wesharewishhair.review.controller.utils.ReviewCreateRequestUtils;
-import lombok.SneakyThrows;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
